requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
dj-pagination==2.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-simpleui==2022.4.9
djangorestframework==3.13.1
fonttools==4.31.2
idna==3.3
install==1.3.5
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
pydicom==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
SimpleITK==2.1.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
